+++ include/psispawn.h
@@ -2,7 +2,7 @@
  * ParaStation
  *
  * Copyright (C) 1999-2003 ParTec AG, Karlsruhe
- * Copyright (C) 2005-2018 ParTec Cluster Competence Center GmbH, Munich
+ * Copyright (C) 2005-2021 ParTec Cluster Competence Center GmbH, Munich
  *
  * This file may be distributed under the terms of the Q Public License
  * as defined in the file LICENSE.QPL included in the packaging of this
@@ -105,7 +105,7 @@ void PSI_registerRankEnvFunc(char **(*func)(int, void *), void *info);
  * @return On success true is returned; or false in case of error
  */
 bool PSI_sendSpawnMsg(PStask_t* task, bool envClone, PSnodes_ID_t dest,
-		      int (*sendFunc)(void *));
+		      ssize_t (*sendFunc)(void *));
 
 /**
  * @brief Send spawn request
